wtc: please look at URL related code, and hooks and nits you might have commented on before.
huanr: please look at sdch_filter code.  The intent was
no semantic change, and only change in histograms and
stats gathered.  I wanted to be sure I had better stats
on several failure cases, as the turn-around time of
adding stats to instrument such cases after they surface
is just too long.

The big feature is the mechanism for getting the total
number of bytes passed to a filter.  We use the filter
context to achieve this, and then the SDCH filter can
calculate compression ratio (from pre-gunzip vs post
SDCH decompress).

The number of bytes read was also histogrammed in a number
of error scenarios, to better diagnose what is going on
when these cases arrise (example: When some data is still
buffered in the VCDIFF decoder).

The sdch_filter destructor was getting long and hard
to read with multiple if blocks, so I cleaned that up
as well a bit (less indentation, and use of early returns).

Nits not included in previous CL that earlier are listed
as well.

r=wtc,huanr
Review URL: http://codereview.chromium.org/40319

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@11665 0039d316-1c4b-4281-b951-d872f2087c98

+++ net/url_request/url_request_job_metrics.cc
@@ -23,8 +23,9 @@ void URLRequestJobMetrics::AppendText(std::wstring* text) {
 
   TimeDelta elapsed = end_time_ - start_time_;
   StringAppendF(text,
-      L"; total bytes read = %d; read calls = %d; time = %lld ms;",
-      total_bytes_read_, number_of_read_IO_, elapsed.InMilliseconds());
+      L"; total bytes read = %ld; read calls = %d; time = %lld ms;",
+      static_cast<long>(total_bytes_read_),
+      number_of_read_IO_, elapsed.InMilliseconds());
 
   if (success_) {
     text->append(L" success.");
