+++ net/url_request/url_request_job_metrics.h
@@ -34,7 +34,7 @@ class URLRequestJobMetrics {
   base::TimeTicks end_time_;
 
   // Total number of bytes the job reads from underline IO.
-  int total_bytes_read_;
+  int64 total_bytes_read_;
 
   // Number of IO read operations the job issues.
   int number_of_read_IO_;
